+++ Assets/CoreAiUnity/Tests/EditMode/LuaProgrammerPipelineEndToEndEditModeTests.cs
@@ -1,3 +1,4 @@
+#if COREAI_HAS_MOONSHARP && !COREAI_NO_LUA
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
@@ -227,3 +228,4 @@ namespace CoreAI.Tests.EditMode
         }
     }
 }
+#endif