+++ Assets/CoreAiUnity/Tests/PlayMode/FastNoLlm/AiGameCommandRouterMainThreadPlayModeTests.cs
@@ -1,3 +1,4 @@
+#if COREAI_HAS_MOONSHARP && !COREAI_NO_LUA
 using System;
 using System.Collections;
 using System.Threading;
@@ -299,3 +300,4 @@ namespace CoreAI.Tests.PlayMode
         }
     }
 }
+#endif